Eliminated redundant menu keyboard listeners
The Kobalte dropdown menus get created and destroyed on each use, so it is
inappropriate to register key event listeners each time!
