+++ online/src/app/classic/components/toolbars.jsx
@@ -1,10 +1,9 @@
 
 import { createSignal, createEffect } from "solid-js";
 
-import { controllerProperty, subController, useEditor } from '../../../viewer/context/editor.jsx';
+import { controllerProperty, subController, useEditor, resumeMenuKeyEvents, suspendMenuKeyEvents } from '../../../viewer/context/editor.jsx';
 import { useSymmetry } from "../context/symmetry.jsx";
 import { ToolConfig } from "../dialogs/toolconfig.jsx";
-import { resumeMenuKeyEvents, suspendMenuKeyEvents } from "../../framework/menus.jsx";
 
 const ToolbarSpacer = () => ( <div style={{ 'min-width': '10px', 'min-height': '10px' }}></div> )
